Consolidate fuzzy matching into main Plagiarism lib
Since we no longer need to differentiate between the two at run time, we can just make fuzzy a super-set of regular matching and simplify the run process into a single comparison.
