+++ services/QuillLMS/engines/evidence/spec/lib/plagiarism_check_spec.rb
@@ -102,6 +102,31 @@ module Evidence
           expect(feedback[:highlight][1][:text]).to eq("From the - passage even - - - though it has --  a ton of spaces")
         end
       end
+
+      context 'fuzzy matching' do
+        let(:feedback) { "this is some standard plagiarism feedback" }
+
+        it 'should identify plagiarism when there is a fuzzy match of 5 or less' do
+          entry = "This phrase plagiarises from the passage even though it has a ton of TYPOspaces in it."
+          passage = "From the passage even though it has a ton of spaces."
+
+          plagiarism_check = Evidence::PlagiarismCheck.new(entry, passage, feedback, rule)
+          feedback = plagiarism_check.feedback_object
+
+          expect(feedback[:optimal]).to be(false)
+        end
+
+        it 'should generate valid entry and passage highlights when fuzzy matching' do
+          entry = "This phrase plagiarises from the passage even though it has a ton of TYPOspaces in it."
+          passage = "From the passage even though it has a ton of spaces."
+
+          plagiarism_check = Evidence::PlagiarismCheck.new(entry, passage, feedback, rule)
+          feedback = plagiarism_check.feedback_object
+
+          expect(entry).to include(feedback[:highlight][0][:text])
+          expect(passage).to include(feedback[:highlight][1][:text])
+        end
+      end
     end
   end
 end
